Added generic data import functions and service demand import
The commercial service demand file has a somewhat unusual structure. First, the technology descriptions are strings enclosed in double quotes, but with commas inside them, which confuses the numpy genfromtxt function. Second, the final two dozen or so lines of the file have an entirely non-conforming structure because they are commentary describing the file contents.

To work around these formatting issues, four somewhat generic file import and supporting functions were added. These functions use the csv module to read the first two lines of a data file and determine the data types and column headings of the file, then re-open the file and read it line-by-line, skipping lines that don't conform to the previously determined data structure.


Former-commit-id: 16e214e062fd4e67583620fdee2e6aff2efcd400