Improved string cleaning function and added test
Data stored as text in the KDBOUT and KSDOUT files does not import as properly formatted text strings. The function str_cleaner, written to remove extraneous spaces and double quote characters from the imported text from KSDOUT, was generalized to handle text from KDBOUT, which has only extraneous spaces, as well. Additionally, a test was created for the function to ensure that it correctly handles both types of text.


Former-commit-id: b66502eae2cda63099d7ad0b4076b9308b03a5d1