Add supply-side effects of demand-side measures
For the heating and cooling end uses, the adoption of demand-side
measures (highly insulating windows, air sealing, etc.) will remove
some of the savings available to supply-side measures (air source heat
pumps, furnaces, etc.), though the two types of measures are not
directly competed. This commit ensures that the secondary effects of
demand-side heating/cooling measure adoptions are accounted for in the
savings potential of supply-side heating/cooling measure adoptions, as
modeled in the ‘compete_measures’ function of ‘run.py’.

Secondary effects of demand-side measures are represented as a ‘demand
adjustment fraction’ that is applied to the initial estimate of a
supply-side measure’s savings, along with the updated market share
calculated by competing this measure against other supply-side measures
accessing the same market microsegment. The demand adjustment fraction
represents the portion of total demand-side energy associated with the
supply-side market microsegment that has NOT already been saved by
demand-side measures.

For example, consider a gas furnace supply-side heating measure that is
being run alongside a highly insulating window demand-side
heating/cooling measure. Let’s say the microsegment key chain and
primary energy use that the supply-side measure accesses is:

•	AIA_CZ1 -> single family homes -> natural gas -> heating -> supply ->
furnace (NG)
•	Represents 100 TBTUs primary energy

The microsegment key chain and primary energy use that the demand-side
measure accesses is:

•	AIA_CZ1 -> single family homes -> natural gas -> heating -> demand ->
windows conduction
•	Represents 500 TBTUs primary energy

The key chain for the total demand-side primary energy associated with
the supply-side microsegment is:

•	AIA_CZ1 -> single family homes -> natural gas -> heating
•	Represents 1000 TBTUs primary energy

The demand adjustment fraction is then the portion of natural gas
heating energy in single family homes in AIA climate zone 1 that is NOT
saved by the highly insulating window measure. In the above example, if
the highly insulating window measure saves 10%, or 50 TBTUs of its 500
TBTU market, the demand adjustment fraction is 1 - (50 TBTUs / 1000
TBTUs total demand side primary energy) = 0.95. Assuming the furnace
measure saves 50%, or 50 TBTUs of its 100 TBTU market, these savings
would then be adjusted to 50 TBTUs x 0.95 = 47.5 TBTU savings,
considering the savings already attributed to the overlapping
demand-side measure. If other supply-side measures were competing
against the furnace measure, an additional market share adjustment
would need to be applied to the furnace measure's savings to account
for the portion of the furnace’s market microsegment that is captured
by these competing measures.

To support the calculation of the demand adjustment fraction, a
‘demand-side adjustment’ key has been added to a measure’s
‘mseg_compete’ attribute.  This key points to a dict that includes the
total demand-side primary energy and demand-side savings information
needed to calculate the demand adjustment fraction associated with a
particular supply-side microsegment. Note that for a supply-side
heating/cooling measure, demand-side savings information is initialized
for each year as zero in the 'mseg_find_partition' function and
subsequently updated as needed in the 'compete_measures' function.

A test been added to 'FindPartitionMasterMicrosegmentTest' in
'run_test.py' to ensure that demand-side adjustment information is
properly updated for a sample supply-side heating/cooling measure, as
well as for a sample water heating measure that has no demand-side
effects (and thus should yield blank demand-side adjustment
information). 'ResCompeteTest' has also been updated to run 2
demand-side and 3 supply-side cooling measures; the test ensures that
any secondary demand-side effects on supply-side measures are
accurately accounted for alongside the updated market shares for each
measure.


Former-commit-id: 7f181857eb085f79260a28487818dbe9245c7ecd