Added handling of service demand data
Added a new function 'sd_mseg_percent' and a corresponding set of tests. The function extracts energy use data from the service demand array for a specific census division, building type, end use, and fuel type and restructures it into a percentage contribution from each technology included in that specific set. While there is a putative 'technology type' column in the service demand data, for some end uses, a single technology type number can correspond to multiple technologies, and the vintage is used as a second identifier for technology type. Unfortunately, unique combinations of technology type and vintage also do not capture all of the records for a single technology. In this function, that problem is overcome by using the Description column text, after being simplified with a regex, to select data by technology. This function is only required for the subset of commercial building end uses that have service demand data.


Former-commit-id: 19a73f830e4095d3a167d07cb4cd24e68a7b03d3