Closes #21
All major commercial microsegment data are now loaded by com_mseg.py. In this commit, code to import the census division to climate zone conversion data and thermal load component data was added. Support for importing the main EIA commercial building data files was completed in commit b66502e.

To support importing the thermal load component data, the custom data import functions were modified to add support for files that are tab-delimited and to have some limited handling of NA entries that appear in R output. The NA missing data placeholder in R does not readily translate to Python, and in the data_import function, they are converted to 'nan'.


Former-commit-id: fe5f36f1d1945e2c8112c3b8f1c08d2dc8698124