Added JSON database interpreter and tests
Added a function that converts a list of keys that define the leaf node in the JSON microsegment database into a list of numbers and strings that can be used to select data in the numpy arrays of commercial data from EIA. This function relies on the dictionaries defined at the top of the com_mseg.py file to do the conversions. There are also now tests corresponding to the new function.


Former-commit-id: 534652523973e30c740435c033582f999eb267c6