Modified recursion to skip some building types
Applying the energy_select function to residential building types will trigger a KeyError. In anticipation of using the function on a complete microsegments JSON with both residential and commercial building types, a line (and explanatory comment) was added to the recursion function 'walk' to call energy_select at the leaf nodes of the database only if the building type key is found in the building type dict.


Former-commit-id: 8b788e78503cb97d97bf69442591c440a03698a9