Fixed division by zero problem in 'sd_mseg_percent'
In the sd_mseg_percent function, the line 'tval = tval/np.sum(tval, axis=0)' raised a RuntimeWarning 'invalid value encountered in true_divide' caused by zeros in the row sum of tval. The problem arose for two cases: 1) when all the service demand data were zero, and 2) when in some years, the sum of all the technology types was zero (notably new england, assembly, cooling, natural gas). In either case, for at least some years, division by zero would be attempted.

Two changes were made to correct this problem. First, if the service demand data are all zero, the division step is skipped and in the energy_select function, the output array is sent to the JSON without any technology-specific information. Second, if the service demand data sum to zero for only some years, the warning is suppressed and the calculation is performed, followed by a step to replace any nan values arising from 0/0 with zeros.

Other approaches to handling the division problem required a try/except block or ran into problems with floating point accuracy.


Former-commit-id: 33f4254024c359629fd8417cd225796d695957c7