Improved compatibility with com. bldgs.; bug fix
In anticipation of a microsegments JSON file that includes both residential and commercial building types, an if statement that checks to make sure the appropriate (residential or commercial) building type is in the list of keys before proceeding with data extraction was added to the 'walk' function.

In mseg.py, the census division to climate zone data were being imported outside the 'main' function. That procedure was moved inside the 'main' function so that when mseg is imported by other routines, a file import action isn't silently triggered. With that change, the ClimConverterTest was failing since it relied on that background file import. A conversion array was written explicitly into ClimConverterTest to ensure that the test can run independently from external files.


Former-commit-id: 1d30a3806d5817c219e51e70c8a22758396e66fc