Fixed problem with &amp; symbol in lighting
One of the lighting types, 'F28T8 HE w/OS & SR', had the ampersand symbol coded as &amp;, which was causing a problem with the data lookup from the JSON file. The symbol was replaced with a '+' in the JSON and in the jQuery text, which resolves the TypeError for this lighting type. A long term fix for this problem will require removing this character in the output from the com_mseg.py routine.


Former-commit-id: aa4cf91cfdf86e99a3f2434719e794e9ce044d4d