Appearance tweaks and updates
Updated CDN reference to the latest version of Bootstrap 3 (3.3.6). Switched to trillion BTU instead of Quads for energy reporting units, and adjusted numbers shown to one decimal place. Removed the 'Recall' and 'Save' buttons since that functionality is not yet implemented and removed the links in the header to other pages that are not completed. Removed disclaimer text now that calculator outputs have been checked against AEO Summary Tables A-4 and A-5. Fixed the problem with the climate zone popover where the image was not appearing in the correct location by preloading the image so that the popover is generated using the appropriate content size information. Closes #92.


Former-commit-id: af9c25f4c8f1a56f603c668e86adf63317b1ff69