Fix comparison of two dicts when testing
As previously written, the tests for the entries in the final data dict were not successfully checking for equality between the expected dict and the output from the function in com_mseg. To fix this problem, a function specifically for comparing two dicts was adopted from tests on the run_scheme branch. Closes #107.


Former-commit-id: 0ded4cd92712fc6242db47ac81256cffad84119e