Updated calculator for accessibility
Earlier changes made the calculator accessible to those who cannot or choose not to use a mouse. These changes focused on improving accessibility for those using screen readers. The changes rely heavily on the Bootstrap 'sr-only' class, which makes content invisible on the page, but still able to be scanned by a screen reader.

For the popover AIA climate zone image, I couldn't get any alt text to be read, so I added .sr-only text before each climate zone button with the definition of the climate zone in heating and cooling degree days. For dropdown menus, the default text was sufficient to explain the action to be taken, but button groups were often less obvious, so I added short phrases as .sr-only before most button groups to explain the next required action. I also updated the jQuery with explanatory .sr-only content for anything added to the DOM as a result of entries on the calculator. Finally, in the Market Size box on the right side of the page, I added further .sr-only text and hid (to the screen reader) some of the text shown visually to help explain the content and provide a clearer statement of the units for those using screen readers.

One lingering problem, which could be caused by VoiceOver, the Boostrap CSS, or incorrect configuration of the DOM, is that all button groups have each button's text read twice, once as a button, and once as the text itself. Closes #37.


Former-commit-id: ba6a6e83c627fb04db9f45a5769f2c4cd51effa5