Revised tech data selection function and tests
The data originally used to set up the tests for the technology data selection function did not include a representative range of technologies that would be extracted by the function from the full EIA data array. The input and output sample data for the tests were modified accordingly. As a result, the cost extraction function now does not work correctly and the test for that function is commented out. That function will be rebuilt and will incorporate service demand data for the average cost, performance, and lifetime calculations.


Former-commit-id: 54c3212e17de013055f60ccc7dc4742ad01589fd