Add service demand weights to technology costs
Several changes were made to be able to incorporate service demand data into the mean or ‘typical’ technology cost calculation.

The functions that subset the service demand or cost, performance, and lifetime data arrays (‘tech_data_selector’ and ‘sd_data_selector’) do so based on microsegment specifications, but the cost/performance/lifetime data extraction functions were planned to operate on a single technology, not all of the technologies in a microsegment. An intermediate function, ‘single_tech_selector,’ was created to take the microsegment level data and extract single technologies for processing by subsequent functions.

Sample arrays of service demand data at various stages of selection and decomposition were added for testing purposes. The test array that represents the full service demand data as imported includes multiple (though not all) technologies for each of the microsegments tested. Including only a single technology in each microsegment was a significant testing design error present in commit #2d1bb62.

The ‘cost_extractor’ function was revised to incorporate service demand data into the calculation of the mean (i.e., ‘typical’) cost. This procedure uses service demand for a single technology normalized for each year as the weights in a weighted arithmetic mean cost calculation. A step was added for the maximum cost calculation to ensure that the maximum is never less than the mean cost for each year.


Former-commit-id: 62988f115c9b48346f59923f9b00f0e1a1d0e1cc