Cost function modified to also extract performance data
Since the process for extracting and preparing the cost and performance data is almost exactly the same, the ‘cost_extractor’ function was renamed ’cost_perf_extractor’ and tweaked to output either parameter depending on the string passed to the ‘flag’ argument. The corresponding test and required sample data for testing were updated accordingly. Comments, variable names, and explanatory docstrings in the function and test were also updated to be more general.


Former-commit-id: cab90df62ef485e432aab5eb90fd1866aa62ff71