Added new relative performance handling and financial outputs
The relative performance of the building stock under the measure vs.
the baseline technology is now calculated separately for the competed
stock and for the previously captured stock in each modeling year. In
the former case, relative performance equals that of the measure vs.
baseline for the current year. In the latter case, relative performance
equals the weighted combination of the relative performances of
captured stocks for all previous years.

Additionally, measure financial metric outputs are added to the
‘write_outputs’ function (e.g., IRR, payback, CCE, CCC) and a filter
that had prevented the calculation of these outputs for scenarios where
measure competition wasn’t run was removed (these outputs may still be
of interest in such cases).

No adjustments to the tests in ‘run_test.py’ are required to
accommodate the above changes.


Former-commit-id: d594c73ed89969e77cbd1506f08fe43e02a091b4