Improve code robustness to large changes in baseline efficiency
This commit addresses an issue where large year-to-year changes in the
efficiency of the baseline technology that a measure is being compared
against were yielding unrealistic fluctuations in the ‘efficient’
energy use of a technical potential scenario over time (see Issue # 115
for more details).

To address this issue, the relative performance of the stock in each
year of a technical potential scenario is calculated as the weighted
sum of the current year’s relative performance and the combined
relative performance of all previous years, where the weights are
assigned as (1 / measure life) and (1 - 1 / measure life), respectively.

These weights represent a minimum annual turnover in the stock,
ensuring the technical potential ‘efficient’ energy use always remains
equal to or lower than that of a maximum adoption potential scenario,
where the stock turnover is more rapid.

Closes # 115


Former-commit-id: b09740295646192b5f686c4f0e6da5b92e9a62a6