Import appropriate EPlus XLSX and initialize measure performance dict
This commit establishes a code base for reading in a series of
efficiency measures, determining which measures require Energy Plus
data for their performance input, and identifying which Energy Plus
simulation output file should be used to fill in the measure
performance information.

The commit also adds a routine, ‘create_perf_dict,’ which initializes
the dictionary that will store the Energy Plus-based performance
information for the measure. The key hierarchy of this dictionary is
determined by the measure’s applicable baseline market information
(climate zones, building types, structure types, fuel types, and end
uses). A new test, ‘CreatePerformanceDictTest,’ has been added to
verify that the ‘create_perf_dict’ routine operates as intended.

Finally, the commit adds a series of dictionaries that will be used to
map the climate zone, building type, fuel type, and end use names used
in Energy Plus simulation output files to the names for these variables
in the Scout analysis framework.


Former-commit-id: f4c0745251cddb2c314dd682f98fe0a692fef485