Add EPlus directory cmd line argument, write updated data to JSON
This commit allows a user to specify a local directory for EnergyPlus
performance files from the command line. The commit also writes any
EnergyPlus-updated measure performance data back to the original
measures JSON database.

In running through the full ‘import_eplus.py’ routine with an example
EnergyPlus performance data XLSX, a few small errors in the code were
discovered. One error required an enhancement to the sample XLSX file
read in by the ‘ConverttoArrayTest’ in ‘import_eplus_test.py’. This
error stems from the inclusion of rows in the EnergyPlus XLSX with
‘Measure Not Applicable’ in the ‘Status’ column, where no energy
savings data are available. Such rows are now removed from the
EnergyPlus data array as part of the ‘convert_to_array’ function in
‘import_eplus.py’.


Former-commit-id: 78a4dd4f12c1073c65ceed61580ff3458849fd24