Function revised to correctly handle specific tech data case
The 'single_tech_selector' function had been incorrectly configured and was removing rows of data corresponding to technologies that did not have performance levels specified with years. This improper removal of rows with the appropriate technology performance data meant empty arrays were being passed to the cost, performance, and lifetime extractor functions for those technologies.

Also, the docstring for the function was revised to Google style standards.


Former-commit-id: 7331f26d0d2b9664ea0efe554d62fbeacd2cf774