Fixed handling of technologies with market exit years before the model period
Added a new test case for the condition where a technology does not have a year in the technology name for any of the reported performance levels. This new test case was intended to test the performance of the already revised 'single_tech_selector' function, but it exposed problems with the cost, performance, and lifetime extraction functions.

The calculation of the 'typical' technology cost and performance in the 'cost_perf_extractor' function was modified so that both are based on the market exit years for each technology performance level. Previously, the 'typical' cost and performance had been based on the service demand data without accounting for the market exit year, which is sometimes earlier than the end of the reported service demand for that technology performance level. The market entry year has not been treated in updating the service demand data, since it appears that there is never service demand reported in years prior to a technology level's market entry year.

Similarly, the 'life_extractor' function was modified to address the specific case where a technology performance level exits the market before the first year of the modeling period.

All of the test data relevant to the new case were updated, and docstrings for the modified functions in com_mseg_tech.py were revised to Google style standards.


Former-commit-id: db3ce8e8c758ed03aea34e0649a4d7bcecff7648