Added function to list technologies in a microsegment
To facilitate the selection and processing of data for a single technology, a function that generates a list of unique technology names for a given microsegment was created. Each technology name in the list can be an input to the 'single_tech_selector' function. Tests for the new function were also added.


Former-commit-id: b38ce00a73f9655efb2092062a7d11778c596a59