Added function to build cost/perf/life dict for a microsegment
Added a function that, for each technology in a given microsegment, calls the various functions that generate appropriately formatted dicts for cost, performance, and lifetime data, and places all of those dicts into a much larger dict for all of the technologies in the microsegment. The function is tested against each of the unique microsegments in the common test cases.


Former-commit-id: a99b1db4cd53fe1ecc340b8bda1efed13a0e4718