Handle ‘add-on’ measure type
Certain measures such as sensors and controls enhance the service of
existing equipment rather than directly replacing this service. Such
measures must be given special handling in the ‘run.py’ routine,
particularly with regard to their installed cost. Specifically, the
installed cost of such ‘add-on’ measures in a new or replacement
adoption decision should be defined as the installed cost of the add-on
measure (as defined in the measures database), plus the installed cost
of the existing equipment to which it is being added.

Measures are therefore now flagged as being either ‘full service’ or
‘add-on’; in the latter case, a baseline equipment cost is added to the
add-on measure’s installed cost as part of the measure’s master
microsegment calculation (see ‘partition_microsegment’ function).

It is noted that certain ‘add-on’ measures would not be expected to
apply to the new and replacement adoption decision types currently
simulated in ‘run.py,’ though they would apply to a retrofit decision
type (not currently simulated). Window films are a good example of this
type of measure: here, a consumer would be more likely to purchase a
window unit with the performance characteristics of a film built-in,
rather than adding the film on top of this new or replacement window
unit.

A measure’s applicability to different decision types will be flagged
in a future commit, once the retrofit decision type has been added.

To support the new ‘add-on’ or ‘full service’ measure type distinction,
a new sample add-on measure has been included in
‘FindPartitionMasterMicrosegmentTest’ in ‘run_test.py’. This new sample
measure is intended to mimic a lighting controls measure that applies
to a commercial building type.

Finally, the ’retrofit’ option for a measure’s structure type attribute
has been renamed to an ‘existing’ structure type option for clarity.


Former-commit-id: d45f24b3ad378f264bd4cb51206e2ccebfb3f509