Consider retrofits as part of measure stock-and-flow calculations
The ‘partition_microsegment’ function in ‘run.py’ now assumes that 2%*
of existing baseline market microsegment stock goes up for retrofit in
each year of the simulation.  This 2% retrofitted stock is added to the
‘competed’ stock, energy, carbon, and cost figures that had previously
been calculated for each measure.

The consideration of retrofitted stock has implications for measure
competition calculations, as a retrofit adoption decision differs from
the new equipment and equipment replacement decisions previously
simulated as part of these competition calculations. A future commit
will adjust measure competition calculations accordingly.

Output values for the ‘FindPartitionMasterMicrosegmentTest’ and
‘PartitionMicrosegmentTest’ tests of ‘run_test.py’ were adjusted to
reflect the new retrofit stock consideration.

* The 2% retrofit rate is based on a 2010 Rocky Mountain Institute
estimate of 2.2%, which applies to commercial buildings (see Olgyay and
Seruto, ‘Whole-Building Retrofits: A Gateway to Climate
Stabilization’). Additional literature suggests the rate of residential
retrofits is less than 2% (see, for example: Neme, Gotten, and
Hamilton, ‘Residential Efficiency Retrofits: A Roadmap for the
Future’).


Former-commit-id: 1c6a75e259d73184e487e27a8728bae3946b4b80