Fixes error on square footage fields
The walk function has been reconfigured to skip cases where there are two or fewer keys (i.e., census division and building type) as in the case of square footage data, since the cost, performance, and lifetime data do not include square footage data. This change fixes an IndexError that occurred in the walk function as a result of referencing a field in the variable mseg_codes that would not exist if only two or fewer keys defined a particular leaf node.


Former-commit-id: 9ba6a7b57f2484ec4f5a0c0e99edab70cb08cc50