Add check of sub-market scaling fraction source information
In cases where sub-market scaling fractions are specified for one or
more of a measure’s applicable baseline market microsegments, the
supporting source information for the scaling fraction(s) is now
checked for validity and completeness.

At a minimum, sub-market scaling fractions must be supported by the
following source information:

1) a valid URL where supporting data resides OR general source
information that can lead one to these data (title, author,
organization, year of a relevant publication), AND
2) a description of how the scaling fraction was derived from the
source data.

Warnings are generated when any of the above information is missing.
When both 1) and 2) are missing, a ‘CRITICAL’ warning notifies the user
that the measure will be removed from the analysis.  The measure’s
status is then changed from ‘active’ to ‘inactive’ and it is removed
from all further routines.

Sub-market source checking is tested by a new ‘test_mseg_warn’ function
within the ‘FindPartitionMasterMicrosegmentTest’ of ‘run_test.py’. This
function runs measures with incomplete sub-market source information
through the ‘mseg_find_partition’ function, ensuring that the proper
warnings are generated and that the measure is marked inactive if
necessary.

Closes #121


Former-commit-id: 3ab06ecf2e49af686ade77a7c3503f99548ada52