Fixes #123
Prior to the changes in this commit, residential and commercial data files were necessarily separate, since each script converted the data from a census division to a climate zone basis immediately prior to exporting the data to a JSON file. Having the partially complete (only residential or commercial) JSON files on a climate zone basis meant that EIA data on a census division basis could no longer be added.

To get residential and commercial data into a single JSON file, the climate conversion step was removed from each individual script and the conversion functions in ‘mseg.py’ were moved to a separate new script. For this new script, the original functions were modified to process combined residential and commercial buildings data and the tests were expanded significantly.

Other changes in the data processing scripts were largely to account for changes to file names and other minor tweaks. Try/except blocks with a more informative message than “FileNotFoundError” were added to the commercial data scripts, which are set up to read in databases with the residential data already added.


Former-commit-id: 8dc7705d977f29bf3239cd909cbc0cbcec4c7807