Improve string handling for cost/perf/life data
The service demand data has technology descriptions that use HTML character encodings for some characters, such as &quot; and &amp; for the double-quote and ampersand characters, respectively. The cost, performance, and lifetime data have these characters as plain text. The code was modified to improve the way that these strings are handled for matching in com_mseg_tech.py. The str_cleaner function borrowed from com_mseg.py was modified to clean up these strings and make them consistent. Since &quot; and -inch, the text used to replace the quote character and HTML encoding string, are different lengths, an adjustment for the string length matching in the cost_perf_extractor function was added.

As part of these changes, tests were added for the particular equipment types where these symbols appear. Creating these tests exposed an example of the concern raised by issue #120, where the calculation of mean and maximum cost does not perform as expected. F28T8 HE w/ OS & SR is a good example of a case where costs are projected to decline over time, and the '2011 typical' performance level for that technology is indicated as exiting the market in 2019, but service demand is given beyond 2019. In 2020, the service demand is used to weight the mean cost and performance, which means that the numbers reported partially reflect a technology that is supposed to have exited the market.

The current approach ensures that the maximum (which would be lower than the mean under normal circumstances since the 2011 typical performance level exits the market in 2019, which is reflected in the val lists within the cost_perf_extractor function) is always at least equal to the mean, but might incorporate performance and cost levels beyond when they should be included in the baseline.


Former-commit-id: e5d8e130e6addb683892b237d90cbc0f64574539