Converted costs to per square foot floor area
To facilitate measure competition, capital costs must be provided with common units. The costs in the EIA data are given on a unit service demand basis, the units of which vary depending on the end use. These data must be converted to a square foot floor area basis using the service demand and square footage data. This conversion is effected by a new function, the output of which is applied within the larger mseg_techology_handler function to the cost data extracted for each technology.

These changes are related to issue #124, but the cost data should be examined further to determine whether the conversion is correct before the issue is closed.


Former-commit-id: 8d892090f94ff4504b6684467d18419df2eae10d