Migrate MeasurePackage subclass and methods from ‘run.py’
Following the previous transfer of the Measure class and associated
methods from ‘run.py’ to ‘measures_prep.py ,’ this commit transfers the
MeasurePackage subclass and associated methods from ‘run.py’ to
‘measures_prep.py’.

Measure packages are assembled after the attributes for all individual
measures that contribute to the package have been finalized via the
‘fill_measures’ function. Once a measure package has been assembled, it
is added to the list of measures exported by ‘measures_prep’ for
further use in the ‘run.py’ analysis engine.

Note that while the ‘merge_measures’ MeasurePackage method is finalized
and tested in this commit, the ‘remove_pkg_overlaps’ method and tests
will be finalized in a subsequent commit.

Finally, Google style Docstrings have been added for all new classes
and functions.


Former-commit-id: 73638f3d8637cdf5b4561ec1c6d7202506991c7e